requirements.txt
attrs==21.2.0
bravado==11.0.3
bravado-core==5.17.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
graphviz==0.16
idna==2.10
joblib==1.0.1
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
kiwisolver==1.3.1
matplotlib==3.4.2
monotonic==1.6
msgpack==1.0.2
neptune-client==0.10.0
numpy==1.21.0
oauthlib==3.1.1
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
psutil==5.8.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
scikit-learn==0.24.2
scipy==1.7.0
simplejson==3.17.3
six==1.16.0
smmap==4.0.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
threadpoolctl==2.2.0
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
torchviz==0.0.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webcolors==1.11.1
websocket-client==1.1.0
